requirements.txt
gspread>=6.0.0
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
pyyaml>=6.0
python-dotenv>=1.0.0
requests>=2.31.0
flask>=3.0.0
trino>=0.328.0
